Init margin component
Several challenges here, because we derive the margin from the line-length, ReadiumCSS not providing with a margin setting in v2, and ts-toolkit takes the pageGutter into account. The issue is that auto layout will add columns if line-length fit, so it only works for an arbitrarily-set number of columns… and the container can effectively contain less that what we request when the multiplier is applied. On the opposite side, pageGutter may not have any effect if the container can fit way more than what we request…